half test

+++ main.py
@@ -35,7 +35,7 @@ def load_model_from_config(config, ckpt, verbose=False):
         print("unexpected keys:")
         print(u)
 
-    model.cuda()
+    model.cuda().half()
     return model
 
 def get_parser(**parser_kwargs):
